+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to
 
 ## Added
 
-✨ Add a custom callout block to the editor #892
+- ✨ Add a custom callout block to the editor #892
+- 🚩(frontend) version MIT only #911
 
 ## [3.2.1] - 2025-05-06